Split BaseLocaleCard component
